+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.8 - (In Development)
 
+1. Fix displaying the endpoints gutter icon when using IntelliJ Community
+   Edition.
 1. Register the XSLT 2.0 and XSLT 3.0 schema file URIs with IntelliJ.
 1. Auto-complete support for Saxon XSLT extension elements.